introduce help and a hack to fetch latest msg

+++ modules/ripaud.py
@@ -10,14 +10,28 @@ ydl_opts = {
     }],
 }
 def ripaud(context):
-    try:
-        os.remove("YTDL/ytdlaudio.mp3")
-    except:
-        pass
-    main.send_msg("starting download!")
-    list = [str(context)]
-    print(list)
-    with youtube_dl.YoutubeDL(ydl_opts) as ydl:
-        ydl.download(list)
-    main.send_msg("download finished, sending audio..")
-    main.send_media("YTDL/ytdlaudio.mp3")
+    if context == "":
+        main.send_msg("No Download Link found")
+    else:
+        try:
+            os.remove("YTDL/ytdlaudio.mp3")
+        except:
+            pass
+        main.send_msg("starting download!")
+        list = [str(context)]
+        print(list)
+        try:
+            with youtube_dl.YoutubeDL(ydl_opts) as ydl:
+                ydl.download(list)
+            main.send_msg("download finished, sending audio..")
+            main.send_media("YTDL/ytdlaudio.mp3")
+        except youtube_dl.utils.DownloadError as e:
+            main.send_msg(str(e))
+
+def help():
+    main.send_msg("""```
+RipVid Usage :
+.ripaud youtubelink
+
+downloads songs and shares in the active chat.
+    ```""")
